Redesign how FrankenTuples are parameterized
Rather than being parameterized on the exact types of the two fields,
they are now parameterized on the type of the tuple field and the names
and type of the named tuple field. This is more convenient in some
cases, as the types of the named elements typically don't matter.